Issue #11163: Split InputHiddenField5 into focused test inputs
